requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
pyjwt==2.12.1
pwdlib==0.3.0
argon2-cffi==25.1.0
email-validator==2.3.0
requests==2.32.5
python-dotenv==1.2.2
streamlit==1.55.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-html==4.2.0
httpx==0.28.1
